+++ CHANGES.rst
@@ -5,7 +5,8 @@ Change log
 5.1 (unreleased)
 ================
 
-- Nothing changed yet.
+- Fix ``SIGCHLD``/``wait`` raise condition associated with the
+  ``start-test-program`` option.
 
 
 5.0 (2023-05-24)